nit: rename `error` -> `errors`
