+++ arbiter-core/src/environment/errors.rs
